+++ src/components/atoms/InlineLink/InlineLink.tsx
@@ -1,6 +1,19 @@
 import { styled } from '@mui/material/styles'
 import Link from 'next/link'
+import { css } from '@mui/system'
 
-export const InlineLink = styled(Link)`
-  padding-left: 4px;
-`
+export const InlineLink = styled(Link, {
+  shouldForwardProp: prop => prop !== 'useHoverUnderline',
+})<{ useHoverUnderline?: boolean }>(
+  ({ useHoverUnderline }) => css`
+    padding-left: 4px;
+    ${useHoverUnderline &&
+    `
+      text-decoration: none;
+      
+      :hover {
+        text-decoration: underline;
+      }
+    `}
+  `,
+)
