--- src/components/atoms/ReadMore/index.ts
@@ -1,1 +1,0 @@
-export { default as ReadMore } from './ReadMore'
