--- src/components/molecules/NewsPostListItem/index.ts
@@ -1,1 +1,0 @@
-export { default as NewsPostListItem } from './NewsPostListItem'
