+++ src/components/molecules/ReadMore/index.ts
@@ -1,0 +1,1 @@
+export { default as ReadMore } from './ReadMore'
