+++ src/components/organisms/NewsAndUpdates/NewsAndUpdates.styled.ts
@@ -1,7 +1,7 @@
 import { styled } from '@mui/material/styles'
 import { css } from '@mui/system'
 import List from '@mui/material/List'
-import Link from 'next/link'
+import { HoverUnderlineLink } from '@atoms/HoverUnderlineLink'
 
 export const NewsPostList = styled(List)(
   () => css`
@@ -13,14 +13,9 @@ export const NewsPostList = styled(List)(
   `,
 )
 
-export const ViewAllNewsLink = styled(Link)`
-  text-decoration: none;
+export const ViewAllNewsLink = styled(HoverUnderlineLink)`
   color: unset;
   display: block;
   width: fit-content;
   margin: 0 auto;
-
-  :hover {
-    text-decoration: underline;
-  }
 `
