+++ src/components/organisms/NewsPostListItem/NewsPostListItem.tsx
@@ -2,7 +2,7 @@ import { ListItemText, Typography, Divider } from '@mui/material'
 import { NewsPostLi, NewsPostAvatar, NewsPostTitleLink, NewsPostImage, NewsPostTitle } from './NewsPostListItem.styled'
 import { NewsPostItemProps } from '@globalTypes/NewsPostItemProps'
 import { getPrettyDateStringFromISOString } from '@helpers/date/date'
-import ReadMore from '@atoms/ReadMore/ReadMore'
+import ReadMore from '@molecules/ReadMore/ReadMore'
 import { getNewsPostImageUrl } from '@helpers/imageUtils'
 
 const NewsPostListItem = (props: NewsPostItemProps) => {
