+++ src/components/organisms/NewsPostListItem/index.ts
@@ -1,0 +1,1 @@
+export { default as NewsPostListItem } from './NewsPostListItem'
