+++ src/theme/styles.css
@@ -15,3 +15,7 @@
 body {
   margin: 0;
 }
+
+a {
+  color: #800080;
+}
